backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
youtube-transcript-api==0.6.0
